Move our code to _ext.go files
